refactor(opal): replace Separator with Divider; simplify InputTypeIn clear API
Separator removal:
- Delete opal Separator component (files + index.ts export)
- Migrate all consumers to Divider: popover, InputSelect, ModelSelector
- InputSelect.Separator compound now aliases Divider directly

Divider vertical fix:
- self-stretch instead of h-full so the line spans the parent flex height
  even when the parent uses items-center
- Remove flex-1 from opal-divider-line-vertical (flex-row context made it
  grow horizontally, fighting w-px); default align-items:stretch handles height

InputTypeIn clearButton API:
- Replace showClearButton + onClear with clearButton?: boolean (default false)
- Default clear fires synthetic onChange("") which works for all consumers
- clearButton is suppressed when rightChildren is set
- Wrappers: InputTypeInField drops onClear interception (onChange handles it);
  InputFile omits clearButton; InputTypeInElementField passes through;
  PasswordInputTypeIn renames prop
- Remove dead onClear calls (set alongside showClearButton={false})
- Migrate all showClearButton → clearButton across codebase
