+++ web/src/app/ee/admin/theme/AppearanceThemeSettings.tsx
@@ -233,7 +233,7 @@ export const AppearanceThemeSettings = forwardRef<
               <InputTypeIn
                 ref={applicationNameInputRef}
                 data-label="application-name-input"
-                showClearButton
+                clearButton
                 variant={errors.application_name ? "error" : undefined}
                 value={values.application_name}
                 {...getPreviewHandlers("sidebar")}
@@ -365,7 +365,7 @@ export const AppearanceThemeSettings = forwardRef<
           <InputTypeIn
             ref={greetingMessageInputRef}
             data-label="greeting-message-input"
-            showClearButton
+            clearButton
             variant={errors.custom_greeting_message ? "error" : undefined}
             value={values.custom_greeting_message}
             {...getPreviewHandlers("greeting")}
@@ -397,7 +397,7 @@ export const AppearanceThemeSettings = forwardRef<
           <InputTypeIn
             ref={headerContentInputRef}
             data-label="chat-header-input"
-            showClearButton
+            clearButton
             variant={errors.custom_header_content ? "error" : undefined}
             value={values.custom_header_content}
             {...getPreviewHandlers("chat_header")}
@@ -467,7 +467,7 @@ export const AppearanceThemeSettings = forwardRef<
               <InputTypeIn
                 ref={customHelpLinkUrlInputRef}
                 data-label="custom-help-link-url-input"
-                showClearButton
+                clearButton
                 placeholder="https://docs.onyx.app"
                 variant={
                   !enterpriseTier
@@ -498,7 +498,7 @@ export const AppearanceThemeSettings = forwardRef<
               <InputTypeIn
                 aria-label="Custom Help Link Label"
                 data-label="custom-help-link-label-input"
-                showClearButton
+                clearButton
                 placeholder="Link label"
                 variant={!enterpriseTier ? "disabled" : undefined}
                 value={values.custom_help_link_label}
@@ -582,7 +582,7 @@ export const AppearanceThemeSettings = forwardRef<
                 <InputTypeIn
                   ref={noticeHeaderInputRef}
                   data-label="notice-header-input"
-                  showClearButton
+                  clearButton
                   variant={errors.custom_popup_header ? "error" : undefined}
                   value={values.custom_popup_header}
                   onChange={(e) =>
