+++ web/src/ee/refresh-pages/admin/HooksPage/HookFormModal.tsx
@@ -121,7 +121,6 @@ function TimeoutField({ spec }: TimeoutFieldProps) {
           type="number"
           placeholder={spec ? String(spec.default_timeout_seconds) : undefined}
           variant={isSubmitting ? "disabled" : undefined}
-          showClearButton={false}
           rightChildren={
             spec?.default_timeout_seconds !== undefined &&
             values.timeout_seconds !== String(spec.default_timeout_seconds) ? (
