+++ web/src/refresh-components/form/InputTypeInField.tsx
@@ -4,10 +4,7 @@ import { useField } from "formik";
 import { InputTypeIn, type InputTypeInProps } from "@opal/components";
 import { useOnChangeEvent, useOnBlurEvent } from "@/hooks/formHooks";
 
-export interface InputTypeInFieldProps extends Omit<
-  InputTypeInProps,
-  "value" | "onClear"
-> {
+export interface InputTypeInFieldProps extends Omit<InputTypeInProps, "value"> {
   name: string;
 }
 
@@ -17,7 +14,7 @@ export default function InputTypeInField({
   onBlur: onBlurProp,
   ...inputProps
 }: InputTypeInFieldProps) {
-  const [field, meta, helpers] = useField(name);
+  const [field, meta] = useField(name);
   const onChange = useOnChangeEvent(name, onChangeProp);
   const onBlur = useOnBlurEvent(name, onBlurProp);
   const hasError = meta.touched && meta.error;
@@ -32,9 +29,6 @@ export default function InputTypeInField({
       value={field.value ?? ""}
       onChange={onChange}
       onBlur={onBlur}
-      onClear={() => {
-        helpers.setValue("");
-      }}
       variant={
         isNonEditable
           ? inputProps.variant
