+++ web/src/refresh-components/popovers/ModelSelector.tsx
@@ -1,15 +1,13 @@
 "use client";
 
 import { useState, useMemo, useRef } from "react";
-import { Popover } from "@opal/components";
 import { LlmManager } from "@/lib/hooks";
 import { getModelIcon } from "@/lib/languageModels";
-import { Button, SelectButton } from "@opal/components";
+import { Button, SelectButton, Popover, Divider } from "@opal/components";
 import { SvgPlusCircle, SvgX } from "@opal/icons";
 import { useSettingsContext } from "@/providers/SettingsProvider";
 import { LLMOption } from "@/refresh-components/popovers/interfaces";
 import ModelListContent from "@/refresh-components/popovers/ModelListContent";
-import { Separator } from "@opal/components";
 
 export const MAX_MODELS = 3;
 
@@ -153,10 +151,10 @@ export default function ModelSelector({
         {selectedModels.length > 0 && (
           <>
             {!atMax && (
-              <Separator
+              <Divider
                 orientation="vertical"
-                paddingXRem={0.5}
-                className="h-5"
+                paddingParallel="sm"
+                paddingPerpendicular="sm"
               />
             )}
             <div className="flex items-center shrink-0">
@@ -176,10 +174,10 @@ export default function ModelSelector({
                     className="flex items-center"
                   >
                     {index > 0 && (
-                      <Separator
+                      <Divider
                         orientation="vertical"
-                        paddingXRem={0.5}
-                        className="h-5"
+                        paddingParallel="sm"
+                        paddingPerpendicular="sm"
                       />
                     )}
                     <SelectButton
