+++ web/src/refresh-pages/admin/ImageGenerationPage/forms/AzureImageGenForm.tsx
@@ -67,7 +67,6 @@ function AzureFormFields(props: ImageGenFormChildProps<AzureFormValues>) {
               <InputTypeIn
                 {...field}
                 placeholder="https://your-resource.cognitiveservices.azure.com/openai/deployments/deployment-name/images/generations?api-version=2025-01-01-preview"
-                showClearButton={false}
                 variant={disabled ? "disabled" : undefined}
               />
             </FormField.Control>
@@ -137,7 +136,6 @@ function AzureFormFields(props: ImageGenFormChildProps<AzureFormValues>) {
                   placeholder={
                     isLoadingCredentials ? "Loading..." : "Enter your API key"
                   }
-                  showClearButton={false}
                   disabled={disabled || !formikProps.values.target_uri?.trim()}
                   error={apiStatus === "error"}
                 />
