+++ web/src/sections/actions/modals/OpenAPIAuthenticationModal.tsx
@@ -433,7 +433,6 @@ export default function OpenAPIAuthenticationModal({
                               value={values.authorizationUrl}
                               onChange={handleChange}
                               placeholder="https://example.com/oauth/authorize"
-                              showClearButton={false}
                             />
                           </FormField.Control>
                           <FormField.Message
@@ -460,7 +459,6 @@ export default function OpenAPIAuthenticationModal({
                               value={values.tokenUrl}
                               onChange={handleChange}
                               placeholder="https://example.com/oauth/access_token"
-                              showClearButton={false}
                             />
                           </FormField.Control>
                           <FormField.Message
@@ -487,7 +485,6 @@ export default function OpenAPIAuthenticationModal({
                               value={values.clientId}
                               onChange={handleChange}
                               placeholder=" "
-                              showClearButton={false}
                             />
                           </FormField.Control>
                           {isEditingOAuthConfig && (
@@ -519,7 +516,6 @@ export default function OpenAPIAuthenticationModal({
                               value={values.clientSecret}
                               onChange={handleChange}
                               placeholder=" "
-                              showClearButton={false}
                             />
                           </FormField.Control>
                           {isEditingOAuthConfig && (
@@ -554,7 +550,6 @@ export default function OpenAPIAuthenticationModal({
                               value={values.scopes}
                               onChange={handleChange}
                               placeholder="e.g. repo, user"
-                              showClearButton={false}
                             />
                           </FormField.Control>
                           <FormField.Description>
