+++ web/src/sections/modals/languageModels/shared.tsx
@@ -636,7 +636,6 @@ export function ModelSelectionField({
                     }
                   }
                 }}
-                showClearButton={false}
               />
             </div>
             <Button
