+++ web/src/sections/sidebar/AdminSidebar.tsx
@@ -258,7 +258,7 @@ function AdminSidebarInner() {
             placeholder="Search..."
             value={query}
             onChange={(e) => setQuery(e.target.value)}
-            showClearButton
+            clearButton
           />
         )}
       </SidebarLayouts.Header>
